[REFACTOR/#89] Alarm 기능 패키지화
