+++ feature/home/src/main/java/com/yapp/alarm/action/AlarmActionContract.kt
@@ -1,4 +1,4 @@
-package com.yapp.alarm
+package com.yapp.alarm.action
 
 import com.yapp.ui.base.UiState
