+++ feature/home/src/main/java/com/yapp/home/HomeNavGraph.kt
@@ -6,8 +6,8 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.navArgument
-import com.yapp.alarm.AlarmActionRoute
-import com.yapp.alarm.AlarmAddEditRoute
+import com.yapp.alarm.action.AlarmActionRoute
+import com.yapp.alarm.addedit.AlarmAddEditRoute
 import com.yapp.common.navigation.OrbitNavigator
 import com.yapp.common.navigation.destination.HomeDestination
